feature: [174] Move classes to other packages
