+++ acl-antlr-dsl/src/main/java/javasabr/mqtt/acl/antlr/dsl/builder/AclRulesBuilder.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.acl.java.dsl.builder;
+package javasabr.mqtt.acl.antlr.dsl.builder;
 
 import java.util.ArrayList;
 import java.util.List;
