+++ acl-antlr-dsl/src/main/java/javasabr/mqtt/acl/antlr/dsl/builder/AllowSubscribeAclRuleBuilder.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.acl.java.dsl.builder;
+package javasabr.mqtt.acl.antlr.dsl.builder;
 
 import javasabr.mqtt.acl.engine.model.condition.TopicCondition;
 import javasabr.mqtt.acl.engine.model.rule.AclRule;
